% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}
}
\value{
a histogram depicting the average sum within a column of data
generated by n and iter and a vector of its values
}
\description{
myclt
}
\examples{
myclt(n = 10, iter = 10000)
}
