% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log likelihood function for data}

\item{x}{data}

\item{param}{parameter values for probability}
}
\value{
a graph depicting the weight of the values including a line where the max likelihood value is
and a list containing the index of the max likelihood value, the parameter value at the max likelihood,
the max likelihood value, and the slopes near the max likelihood.
}
\description{
mymaxlik
}
\examples{
lfunbef = function(y, p) {
n = 20  # Number of trials in each binomial experiment
sum(y * log(p) + (n - y) * log(1 - p))
}

lfun = Vectorize(lfunbef)

y = c(3, 3, 4, 3, 4, 5, 5, 4)

p_values = seq(0.01, 0.99, by = 0.01)

mymaxlik(lfun, y, p_values)
}
