% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{target overbooking probability}

\item{p}{probability a passenger shows up}
}
\value{
Two plots demonstrating likelihood of passengers showing up based on variables
}
\description{
ntickets
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
